Fix drag and mostly fix terminal velocity.
Known issue: terminal velocity is reached and ship cannot be redirected unless key is released.

+++ AsteroidsThruster/src/Main.java
@@ -155,10 +155,10 @@ public class Main implements Runnable, KeyListener {
         } else {
             g.setColor(Color.white);
             g.drawString("GAME OVER", 450, 600);
-            g.drawImage(explosion, spaceship.xpos, spaceship.ypos, 75, 75, null);
-            g.drawImage(explosion, spaceship.xpos, spaceship.ypos, 75, 75, null);
-            g.drawImage(explosion, spaceship.xpos, spaceship.ypos, 75, 75, null);
-            g.drawImage(explosion, spaceship.xpos, spaceship.ypos, 75, 75, null);
+            g.drawImage(explosion, (int)spaceship.xpos, (int)spaceship.ypos, 75, 75, null);
+            g.drawImage(explosion, (int)spaceship.xpos, (int)spaceship.ypos, 75, 75, null);
+            g.drawImage(explosion, (int)spaceship.xpos, (int)spaceship.ypos, 75, 75, null);
+            g.drawImage(explosion, (int)spaceship.xpos, (int)spaceship.ypos, 75, 75, null);
         }
 
         g.dispose();
